requirements.txt
# Core deep learning framework
torch>=2.0.0
torchvision>=0.15.0

# TPU support (optional)
torch-xla>=2.0.0; platform_system != "Windows"

# Image processing
opencv-python>=4.8.0
Pillow>=10.0.0
scikit-image>=0.21.0

# Numerical computing
numpy>=1.24.0
scipy>=1.11.0

# Data handling and utilities
matplotlib>=3.7.0
tqdm>=4.65.0
pyyaml>=6.0
requests>=2.31.0
psutil>=5.9.0

# Configuration and logging
hydra-core>=1.3.0
wandb>=0.15.0
tensorboard>=2.13.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Development tools
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0